Fixed returns not matching in Post
Some returns of the getters didn't match types with the attributes

+++ src/main/java/Post.java
@@ -8,7 +8,7 @@ public class Post {
     private List<String> liked_by;
     private List<String> commented_by;
 
-    public float getId() {
+    public Integer getId() {
         return id;
     }
